Add libasound2-dev

+++ setup.sh
@@ -9,6 +9,12 @@ if [[ -z "$1" ]] ; then
   exit 1
 fi
 
+libasound_status="$(dpkg-query -W -f='${Status}' libasound2-dev 2> /dev/null || true)"
+if [[ "$libasound_status" != "install ok installed" ]] ; then
+  echo "Missing required package: libasound2-dev"
+  exit 1
+fi
+
 rm -rf tmp lib/3rd_party/portaudio lib/3rd_party/onnxruntime
 
 mkdir -p tmp/portaudio
